+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # QOwnNotes Changelog
 
+## 18.03.2
+- there now is a new scripting function `script.insertMedia()` to insert a media
+  file into the media folder (for [#869](https://github.com/pbek/QOwnNotes/issues/869))
+    - for more information please take a look at the [scripting documentation](http://docs.qownnotes.org/en/develop/scripting/README.html#inserting-a-media-file-into-the-media-folder)
+- there now is a new script `Scribble` in the *Scripting repository* to insert a
+  scribble image to the media-folder, that will be edited by an external 
+  image manipulation application (for [#869](https://github.com/pbek/QOwnNotes/issues/869))
+- switched from SourceForge to TuxFamily for the hosting of the sourcecode 
+  archives for the releases of AUR, Snap, Slackware and Gentoo
+    - you will find the archives on [QOwnNotes archives](https://download.tuxfamily.org/qownnotes/src)
+
 ## 18.03.1
 - fixed translation files